# Improve performance of parsing