@sonnyp 2018-09-29
parsers suite:

before: `ltx x 817,222 ops/sec ±1.94% (88 runs sampled)`
after: `ltx x 827,444 ops/sec ±1.26% (93 runs sampled)`

Node.js v10.11.0 - Intel(R) Core(TM) i5-2520M CPU @ 2.50GHz