@sonnyp 2018-09-29
@mogsie this looks good but the benchmark does not report a significant increase in performance. Do you have data that suggests otherwise?