@mogsie 2018-09-29
Hmm, that's interesting.  It obviously only has an effect on the test if
* There is a fair amount of text (including whitespace) between the tags
* There are any comments
* There are attributes with relatively large values (e.g. 100 characters, typially URLs)

Then again, I did my initial tests on node 6 and node 8, nodejs might have optimized their way out of the problem, but if the document you're testing is `<foo bar="baz">quux</foo>` then it's not a surprise that there is no change, as it has no comments, almost no text and attributes. 

I tried replacing the document with [the example atom document from Wikipedia](https://en.wikipedia.org/wiki/Atom_%28Web_standard%29#Example_of_an_Atom_1.0_feed):

before: `ltx x 28,432 ops/sec ±0.83% (91 runs sampled)`
after: `ltx x 33,382 ops/sec ±0.86% (89 runs sampled)`

after:
```
suite XML parsers comparison
sax x 8,447 ops/sec ±1.25% (92 runs sampled)
node-xml x 7,407 ops/sec ±0.59% (89 runs sampled)
libxmljs x 7,865 ops/sec ±0.65% (89 runs sampled)
node-expat x 9,205 ops/sec ±0.49% (92 runs sampled)
ltx x 33,382 ops/sec ±0.86% (89 runs sampled)
```

Still more than three times as fast as node-expat.