@mogsie 2018-09-30
I made a simpler change to show how `master` doesn't scale well with longer texts.

master, unchanged, on my machine (i7-2670QM @ 2.20GHz) w/node 9.11.2
```
suite XML parsers comparison
sax x 158,052 ops/sec ±1.39% (85 runs sampled)
node-xml x 130,975 ops/sec ±0.70% (91 runs sampled)
libxmljs x 131,653 ops/sec ±0.73% (93 runs sampled)
node-expat x 227,446 ops/sec ±0.42% (92 runs sampled)
ltx x 538,775 ops/sec ±0.99% (91 runs sampled)
Fastest is "ltx"
```

If I change the benchmark as follows, just lengthening the attribute and text value, the characteristics change:

```
diff --git a/benchmarks/parsers.js b/benchmarks/parsers.js
index 4ebee36..0a21b8a 100644
--- a/benchmarks/parsers.js
+++ b/benchmarks/parsers.js
@@ -64,7 +64,7 @@ var suite = new benchmark.Suite('XML parsers comparison')
 parsers.forEach(function (parser) {
   parser.parse('<r>')
   suite.add(parser.name, function () {
-    parser.parse('<foo bar="baz">quux</foo>')
+    parser.parse('<foo bar="urn:uuid:60a76c80-d399-11d9-b91C-0003939e0af6">urn:uuid:60a76c80-d399-11d9-b91C-0003939e0af6 urn:uuid:60a76c80-d399-11d9-b91C-0003939e0af6</foo>')
   })
 })

```

```
suite XML parsers comparison
sax x 79,411 ops/sec ±0.58% (93 runs sampled)
node-xml x 106,060 ops/sec ±0.88% (90 runs sampled)
libxmljs x 118,203 ops/sec ±0.42% (94 runs sampled)
node-expat x 179,483 ops/sec ±0.37% (94 runs sampled)
ltx x 283,425 ops/sec ±0.84% (92 runs sampled)
Fastest is "ltx"
```

All tests slow down, some more significantly than others.  ltx from 538k to 283k.  This PR causes the slowdown to go from 538k to 489k

* original
  `ltx x 538,775 ops/sec ±0.99% (91 runs sampled)`
* master w/ longer texts
  `ltx x 283,425 ops/sec ±0.84% (92 runs sampled)`
* pr 126 w/longer texts
  `ltx x 489,294 ops/sec ±0.37% (89 runs sampled)`

Do you want me to add the longer texts to the benchmark in this PR?