@sonnyp 2018-09-30
> Do you want me to add the longer texts to the benchmark in this PR?

Yes please.

:+1: good job!