@sonnyp 2018-10-03
https://github.com/xmppjs/ltx/releases/tag/v2.8.0